Possible new version
Encapsulates everything in a single loop. The additional loop for
applying gzip was left out for sake of clarity. The code will work fine
on Mac OS, but it may skip the first folder in other systems that do
have the .DS_Store folder. The code works for images of any size
smaller than 256x256, but keep in mind images of size 28x28 are
required for MNIST-based machine learning algorithms.